+++ Source/IGListAdapter.h
@@ -96,11 +96,22 @@ IGLK_SUBCLASSING_RESTRICTED
  the previous and succeeding 2 objects will be notified that they are within the working range. As you scroll the list
  the range is updated as objects enter and exit the working range.
 
- To opt out of using the working range, you can provide a value of `0`.
+ To opt out of using the working range, use `initWithUpdater:viewController:` or provide a working range of `0`.
  */
 - (instancetype)initWithUpdater:(id <IGListUpdatingDelegate>)updater
                  viewController:(nullable UIViewController *)viewController
                workingRangeSize:(NSInteger)workingRangeSize NS_DESIGNATED_INITIALIZER;
+
+/**
+ Initializes a new `IGListAdapter` object with a working range of `0`.
+ 
+ @param updater An object that manages updates to the collection view.
+ @param viewController The view controller that will house the adapter.
+ 
+ @return A new list adapter object.
+ */
+- (instancetype)initWithUpdater:(id <IGListUpdatingDelegate>)updater
+                 viewController:(nullable UIViewController *)viewController;
 
 /**
  Perform an update from the previous state of the data source. This is analogous to calling