+++ Tests/IGListAdapterTests.m
@@ -180,8 +180,7 @@
     self.dataSource.objects = @[@0, @1, @2];
     UIViewController *controller = [UIViewController new];
     IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:[IGListReloadDataUpdater new]
-                                                     viewController:controller
-                                                   workingRangeSize:0];
+                                                     viewController:controller];
     adapter.collectionView = self.collectionView;
     adapter.dataSource = self.dataSource;
     IGListSectionController *sectionController = [adapter sectionControllerForObject:@1];
@@ -191,8 +190,7 @@
 - (void)test_whenSettingCollectionView_thenSettingDataSource_thatCellExists {
     self.dataSource.objects = @[@1];
     IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:[IGListReloadDataUpdater new]
-                                                     viewController:nil
-                                                   workingRangeSize:0];
+                                                     viewController:nil];
     adapter.collectionView = self.collectionView;
     adapter.dataSource = self.dataSource;
     [self.collectionView layoutIfNeeded];
@@ -202,8 +200,7 @@
 - (void)test_whenSettingDataSource_thenSettingCollectionView_thatCellExists {
     self.dataSource.objects = @[@1];
     IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:[IGListReloadDataUpdater new]
-                                                     viewController:nil
-                                                   workingRangeSize:0];
+                                                     viewController:nil];
     adapter.dataSource = self.dataSource;
     adapter.collectionView = self.collectionView;
     [self.collectionView layoutIfNeeded];
@@ -213,7 +210,7 @@
 - (void)test_whenChangingCollectionViews_thatCellsExist {
     self.dataSource.objects = @[@1];
     IGListAdapterUpdater *updater = [[IGListAdapterUpdater alloc] init];
-    IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:updater viewController:nil workingRangeSize:0];
+    IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:updater viewController:nil];
     adapter.dataSource = self.dataSource;
     adapter.collectionView = self.collectionView;
     [self.collectionView layoutIfNeeded];
@@ -229,13 +226,13 @@
     IGListTestAdapterDataSource *dataSource1 = [[IGListTestAdapterDataSource alloc] init];
     dataSource1.objects = @[@1];
     IGListAdapterUpdater *updater1 = [[IGListAdapterUpdater alloc] init];
-    IGListAdapter *adapter1 = [[IGListAdapter alloc] initWithUpdater:updater1 viewController:nil workingRangeSize:0];
+    IGListAdapter *adapter1 = [[IGListAdapter alloc] initWithUpdater:updater1 viewController:nil];
     adapter1.dataSource = dataSource1;
 
     IGListTestAdapterDataSource *dataSource2 = [[IGListTestAdapterDataSource alloc] init];
     dataSource1.objects = @[@1];
     IGListAdapterUpdater *updater2 = [[IGListAdapterUpdater alloc] init];
-    IGListAdapter *adapter2 = [[IGListAdapter alloc] initWithUpdater:updater2 viewController:nil workingRangeSize:0];
+    IGListAdapter *adapter2 = [[IGListAdapter alloc] initWithUpdater:updater2 viewController:nil];
     adapter1.dataSource = dataSource2;
 
     // associate collection view with adapter1
@@ -476,7 +473,7 @@
         dataSource.objects = @[@0, @1, @2];
 
         IGListReloadDataUpdater *updater = [[IGListReloadDataUpdater alloc] init];
-        IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:updater viewController:nil workingRangeSize:0];
+        IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:updater viewController:nil];
         adapter.collectionView = collectionView;
         adapter.dataSource = dataSource;
         weakAdapter = adapter;
@@ -516,7 +513,7 @@
         dataSource.objects = @[@0, @1, @2];
 
         IGListReloadDataUpdater *updater = [[IGListReloadDataUpdater alloc] init];
-        IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:updater viewController:nil workingRangeSize:0];
+        IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:updater viewController:nil];
         adapter.collectionView = collectionView;
         adapter.dataSource = dataSource;
         weakAdapter = adapter;
@@ -756,8 +753,7 @@
 - (void)test_whenReloadingData_withNoDataSource_thatCompletionCalledWithNO {
     self.dataSource.objects = @[@1];
     IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:[IGListReloadDataUpdater new]
-                                                     viewController:nil
-                                                   workingRangeSize:0];
+                                                     viewController:nil];
     adapter.collectionView = self.collectionView;
 
     __block BOOL executed = NO;
@@ -771,8 +767,7 @@
 - (void)test_whenReloadingData_withNoCollectionView_thatCompletionCalledWithNO {
     self.dataSource.objects = @[@1];
     IGListAdapter *adapter = [[IGListAdapter alloc] initWithUpdater:[IGListReloadDataUpdater new]
-                                                     viewController:nil
-                                                   workingRangeSize:0];
+                                                     viewController:nil];
     adapter.dataSource = self.dataSource;
 
     __block BOOL executed = NO;